move to separate file and improve api
